fixed helper to make react optional

+++ scripts/helper.js
@@ -383,54 +383,59 @@ HELPER.precompile = (() => {
 		 * Compile and autoprefix Sass
 		 */
 		reactSass: () => {
-			// 1. create directory
-			CreateDir('./lib/css/');
+			if( Fs.existsSync(`${ process.cwd() }/src/css/styles.scss`) ) {
 
-			// 2. compile scss
-			Sassify('./src/css/styles.scss', './lib/css/styles.css');
+				// 1. create directory
+				CreateDir('./lib/css/');
 
-			// 3. autoprefixer
-			Autoprefix('./lib/css/styles.css');
+				// 2. compile scss
+				Sassify('./src/css/styles.scss', './lib/css/styles.css');
+
+				// 3. autoprefixer
+				Autoprefix('./lib/css/styles.css');
+			}
 		},
 
 		/**
 		 * Transpile react to es5, compile css file and include it into our react component
 		 */
 		react: () => {
-			const reactOptions = {
-				ast: false,
-				compact: true,
-				minified: true,
-				presets: [
-					`es2015`,
-					`react`,
-					`stage-0`,
-				],
-				sourceMaps: "both",
-				sourceMapTarget: `react.es5.js`,
-			};
+			if( Fs.existsSync(`${ process.cwd() }/lib/js/react.js`) ) {
+				const reactOptions = {
+					ast: false,
+					compact: true,
+					minified: true,
+					presets: [
+						`es2015`,
+						`react`,
+						`stage-0`,
+					],
+					sourceMaps: "both",
+					sourceMapTarget: `react.es5.js`,
+				};
 
-			const searches = {
-				'/* [replace-includes] */': `import '../css/styles.css';`,
-			};
+				const searches = {
+					'/* [replace-includes] */': `import '../css/styles.css';`,
+				};
 
-			// 1. Copy files
-			CopyFile('./lib/js/react.js', './lib/js/react.es5.js');
+				// 1. Copy files
+				CopyFile('./lib/js/react.js', './lib/js/react.es5.js');
 
-			// 2. Replace the comment with an import statement
-			ReplaceFileContent( searches, `${ process.cwd() }/lib/js/react.es5.js` );
+				// 2. Replace the comment with an import statement
+				ReplaceFileContent( searches, `${ process.cwd() }/lib/js/react.es5.js` );
 
-			// 1. Compile /lib/react.js to react.es5.js
-			Babel.transformFile( `./lib/js/react.es5.js`, reactOptions, ( error, result ) => {
-				if( error ) {
-					HELPER.log.error(`We encountered an error when transpiling the react file in ${ Chalk.yellow( `${ process.cwd() }/lib/js/react.es5.js` ) }`);
-					HELPER.log.error( error );
-				}
-				else {
-					Fs.writeFileSync( `./lib/js/react.es5.js`, result.code );
-					Fs.writeFileSync( `./lib/js/react.es5.js.map`, JSON.stringify( result.map, null, 2 ) );
-				}
-			});
+				// 1. Compile /lib/react.js to react.es5.js
+				Babel.transformFile( `./lib/js/react.es5.js`, reactOptions, ( error, result ) => {
+					if( error ) {
+						HELPER.log.error(`We encountered an error when transpiling the react file in ${ Chalk.yellow( `${ process.cwd() }/lib/js/react.es5.js` ) }`);
+						HELPER.log.error( error );
+					}
+					else {
+						Fs.writeFileSync( `./lib/js/react.es5.js`, result.code );
+						Fs.writeFileSync( `./lib/js/react.es5.js.map`, JSON.stringify( result.map, null, 2 ) );
+					}
+				});
+			}
 		},
 	}
